requirements.txt
tweepy==4.14.0
openai==0.28.1
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiosmtplib==2.0.2
jinja2==3.1.2
python-jose==3.3.0
flask==3.0.0
gunicorn==21.2.0
watchdog==3.0.0
